gpt_backend/chat_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pytest
httpx
pyyaml
requests
structlog
python-json-logger
pymongo
pyjwt
